projetosPython/Projeto_de_Portfolio/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.40
botocore==1.34.40
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.9
django-appconf==1.0.6
django-compressor==4.4
django-heroku==0.3.1
django-htmlmin==0.11.0
django-storages==1.14.2
gunicorn==21.2.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.4
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
webencodings==0.5.1
whitenoise==6.6.0
